backend/requirements.txt
# TruthGuard Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP Client (async)
httpx==0.27.2

# Config & Environment
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Search & Fact-Checking
tavily-python==0.5.0

# Utilities
python-multipart==0.0.12
